Reformats send_updates
Breaks logic out into separate functions
Changes behaviour:

Old behaviour -
        1) Get latest spoilers
        2) Get all users
        3) For each user, for each spoiler...
          1) Upload the spoiler
          2) Get the attachment ID
          3) Send the spoiler to the user
          4) Save the spoiler to the database

New behaviour -
        1) Get latest spoilers
        2) Upload all the spoilers, map attach_id => spoiler w/ dict
        3) Get all users
        4) For each user, for each spoiler...
                1) Send the spoiler to the user using the dict
        5) Save the spoilers to the database

This is a slight speed improvement for each user, but if the server
crashes during this process it can result in users getting many repeat
submissions.

Recommendation is to add an autoincrement ID to the spoilers table and a
last_spoiled ID to each user. This means that spoilers can be saved to
the database *before* sending them out to any users, as the current ID
can be compared with the users table to see who may be missing spoilers.